+++ packages/libraries/main/src/matcher/dictionary/feedback.ts
@@ -1,4 +1,4 @@
-import Options from '../../Options'
+import zxcvbnOptions from '../../Options'
 import { MatchEstimated } from '../../types'
 import { ALL_UPPER_INVERTED, START_UPPER } from '../../data/const'
 
@@ -9,14 +9,14 @@ const getDictionaryWarningPassword = (
   let warning = ''
   if (isSoleMatch && !match.l33t && !match.reversed) {
     if (match.rank <= 10) {
-      warning = Options.translations.warnings.topTen
+      warning = zxcvbnOptions.translations.warnings.topTen
     } else if (match.rank <= 100) {
-      warning = Options.translations.warnings.topHundred
+      warning = zxcvbnOptions.translations.warnings.topHundred
     } else {
-      warning = Options.translations.warnings.common
+      warning = zxcvbnOptions.translations.warnings.common
     }
   } else if (match.guessesLog10 <= 4) {
-    warning = Options.translations.warnings.similarToCommon
+    warning = zxcvbnOptions.translations.warnings.similarToCommon
   }
   return warning
 }
@@ -27,7 +27,7 @@ const getDictionaryWarningWikipedia = (
 ) => {
   let warning = ''
   if (isSoleMatch) {
-    warning = Options.translations.warnings.wordByItself
+    warning = zxcvbnOptions.translations.warnings.wordByItself
   }
   return warning
 }
@@ -37,9 +37,9 @@ const getDictionaryWarningNames = (
   isSoleMatch?: Boolean,
 ) => {
   if (isSoleMatch) {
-    return Options.translations.warnings.namesByThemselves
+    return zxcvbnOptions.translations.warnings.namesByThemselves
   }
-  return Options.translations.warnings.commonNames
+  return zxcvbnOptions.translations.warnings.commonNames
 }
 
 const getDictionaryWarning = (match: MatchEstimated, isSoleMatch?: Boolean) => {
@@ -54,7 +54,7 @@ const getDictionaryWarning = (match: MatchEstimated, isSoleMatch?: Boolean) => {
   } else if (isAName) {
     warning = getDictionaryWarningNames(match, isSoleMatch)
   } else if (dictName === 'userInputs') {
-    warning = Options.translations.warnings.userInputs
+    warning = zxcvbnOptions.translations.warnings.userInputs
   }
   return warning
 }
@@ -65,15 +65,15 @@ export default (match: MatchEstimated, isSoleMatch?: Boolean) => {
   const word = match.token
 
   if (word.match(START_UPPER)) {
-    suggestions.push(Options.translations.suggestions.capitalization)
+    suggestions.push(zxcvbnOptions.translations.suggestions.capitalization)
   } else if (word.match(ALL_UPPER_INVERTED) && word.toLowerCase() !== word) {
-    suggestions.push(Options.translations.suggestions.allUppercase)
+    suggestions.push(zxcvbnOptions.translations.suggestions.allUppercase)
   }
   if (match.reversed && match.token.length >= 4) {
-    suggestions.push(Options.translations.suggestions.reverseWords)
+    suggestions.push(zxcvbnOptions.translations.suggestions.reverseWords)
   }
   if (match.l33t) {
-    suggestions.push(Options.translations.suggestions.l33t)
+    suggestions.push(zxcvbnOptions.translations.suggestions.l33t)
   }
   return {
     warning,